chore: Migrate util tests to typescript (#14548)
